src/publisher/requirements.txt
opentelemetry-api==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-exporter-otlp==1.17.0
six==1.15.0
thrift==0.13.0
